chore: rename Services to Store
